240208 1743 controller, dto, entity, repository, service 수정, selectitem_list.html삭제, item_list.html, bom_list.html 수정, 대문자/숫자만 입력 가능한 필드

+++ sixbeam/src/main/java/com/erpproject/sixbeam/pd/dto/OrderDto.java
@@ -13,7 +13,7 @@ public class OrderDto {
     private Date orderDelivDt;
     private EmpInfoEntity empInfoEntity;
     private ItemEntity itemEntity;
-    private int orderAmt;
+    private Long orderAmt;
     private boolean orderSt;
 
     public OrderEntity toEntity() {
